gistops/confluence/requirements.txt
jsonschema==4.16.0
jinja2==3.1.2
invoke==1.7.3
fire==0.4.0
pyyaml==6.0
semver==2.13.0
atlassian-python-api==3.28.1